chore: style fixes and code improvements
